+++ package.json
@@ -13,7 +13,7 @@
     "lint:fix": "eslint --fix",
     "unit": "c8 --100 node --test",
     "test": "npm run unit && npm run test:typescript",
-    "test:typescript": "tsd",
+    "test:typescript": "tstyche",
     "ajv:compile": "ajv compile -s test/source.json -o test/validate_schema.js"
   },
   "repository": {
@@ -74,7 +74,7 @@
     "neostandard": "^0.13.0",
     "require-from-string": "^2.0.2",
     "sanitize-filename": "^1.6.3",
-    "tsd": "^0.33.0"
+    "tstyche": "^7.0.0"
   },
   "dependencies": {
     "ajv": "^8.12.0",
